+++ src/engines/ml/deepl.ts
@@ -5,6 +5,7 @@ import {
     TranslateOptions
 } from "$/type";
 import {flattenObject, unflattenObject} from "$/util";
+import {defaultLogger} from "$/logger";
 
 interface DeepLTranslationEntry {
     detected_source_language: string,
@@ -65,8 +66,10 @@ export function createDeepLTranslateEngine(config: DeepLConfig): TranslateEngine
 
             const languagesTranslations: any = {}
 
+            const logger = options.logger || defaultLogger;
+
             for (let targetLanguageCode of options.targetLanguageCodes) {
-                if (options.debug) console.log(`DeepL translate > Translating '${options.baseLanguageCode}'>'${targetLanguageCode}'...`);
+                logger.engineDebug('DeepL', `Translating '${options.baseLanguageCode}' > '${targetLanguageCode}'`);
                 const baseTranslationsValues = Object.values(flatTranslations);
 
                 const result = await fetchTranslations(targetLanguageCode, baseTranslationsValues, options);
@@ -95,8 +98,10 @@ export function createDeepLTranslateEngine(config: DeepLConfig): TranslateEngine
 
             const languagesTranslations: any = {}
 
+            const logger = options.logger || defaultLogger;
+
             for (let targetLanguageCode in missingTranslations.targetLanguageTranslationsKeys) {
-                if (options.debug) console.log(`DeepL translate > Translating '${options.baseLanguageCode}'>'${targetLanguageCode}'...`);
+                logger.engineDebug('DeepL', `Translating '${options.baseLanguageCode}' > '${targetLanguageCode}'`);
                 const missingKeys = Object.keys(flattenObject(missingTranslations.targetLanguageTranslationsKeys[targetLanguageCode]));
 
                 const missingValues: string[] = []