+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.6.0] - 2025-08-01
 
 - Added comprehensive logging system with debug/verbose levels
+- Added total execution time
 - Added tests
 - Replaced logWithColor with new logger interface